requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bandit==1.7.4
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
keyring==23.5.0
lazy-object-proxy==1.7.1
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy==0.931
mypy-extensions==0.4.3
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2022.1.8
pkginfo==1.8.2
platformdirs==2.5.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pylama==8.3.7
pylint==2.12.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
readme-renderer==33.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rope==0.23.0
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sqlparse==0.4.2
stack-data==0.2.0
stevedore==3.5.0
terminado==0.13.2
testpath==0.6.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
twine==3.8.0
typing_extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
zipp==3.7.0
